refactor: consolidate file domains under global/file package
- Move all file domain classes from individual domains to global/file/domain
- Relocate FileDimension from global/model to global/file
- Update all import statements to reflect new package structure
- Fix import ordering and star imports to comply with checkstyle
