+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/auth/service/KakaoValidateService.java
@@ -5,7 +5,6 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URI;
-import java.net.URL;
 import org.json.JSONObject;
 import org.springframework.stereotype.Service;
