+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/content/endpoint/request/StoryGalleryWriteRequest.java
@@ -1,10 +1,11 @@
 package io.itmca.lifepuzzle.domain.content.endpoint.request;
 
 import io.itmca.lifepuzzle.domain.content.entity.Story;
+import lombok.Getter;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
-import lombok.Getter;
 
 @Getter
 public class StoryGalleryWriteRequest {
