+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/content/entity/Gallery.java
@@ -1,35 +1,21 @@
 package io.itmca.lifepuzzle.domain.content.entity;
 
-import static io.itmca.lifepuzzle.global.constants.FileConstant.ORIGINAL_BASE_PATH;
-import static io.itmca.lifepuzzle.global.constants.ServerConstant.S3_SERVER_HOST;
-
 import io.itmca.lifepuzzle.domain.content.type.AgeGroup;
 import io.itmca.lifepuzzle.domain.content.type.GalleryType;
 import io.itmca.lifepuzzle.global.file.CustomFile;
 import io.itmca.lifepuzzle.global.jpa.converter.JsonListConverter;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Convert;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Lob;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import jakarta.persistence.*;
+import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import org.springframework.util.StringUtils;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+import static io.itmca.lifepuzzle.global.constants.FileConstant.ORIGINAL_BASE_PATH;
+import static io.itmca.lifepuzzle.global.constants.ServerConstant.S3_SERVER_HOST;
 
 @Table(name = "story_photo")
 @Entity
