+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/content/entity/StoryGalleryMapId.java
@@ -1,11 +1,8 @@
 package io.itmca.lifepuzzle.domain.content.entity;
 
+import lombok.*;
+
 import java.io.Serializable;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
