+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/content/repository/GalleryRepository.java
@@ -1,10 +1,11 @@
 package io.itmca.lifepuzzle.domain.content.repository;
 
 import io.itmca.lifepuzzle.domain.content.entity.Gallery;
-import java.util.List;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface GalleryRepository extends JpaRepository<Gallery, Long> {
