+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/content/repository/LikeRepository.java
@@ -2,10 +2,11 @@ package io.itmca.lifepuzzle.domain.content.repository;
 
 import io.itmca.lifepuzzle.domain.content.entity.Like;
 import io.itmca.lifepuzzle.domain.content.type.LikeType;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+
+import java.util.Optional;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
   @Query("SELECT l FROM Like l WHERE l.id = :userId AND l.targetId = :targetId AND l.type = :type")
