+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/content/repository/StoryRepository.java
@@ -1,12 +1,13 @@
 package io.itmca.lifepuzzle.domain.content.repository;
 
 import io.itmca.lifepuzzle.domain.content.entity.Story;
-import java.util.List;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface StoryRepository extends JpaRepository<Story, String> {
