+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/user/service/UserWriteService.java
@@ -9,12 +9,12 @@ import io.itmca.lifepuzzle.domain.hero.type.HeroAuthStatus;
 import io.itmca.lifepuzzle.domain.user.endpoint.request.UserUpdateRequest;
 import io.itmca.lifepuzzle.domain.user.entity.User;
 import io.itmca.lifepuzzle.domain.user.entity.UserHeroShare;
-import io.itmca.lifepuzzle.domain.user.file.UserProfileImage;
 import io.itmca.lifepuzzle.domain.user.repository.UserHeroShareRepository;
 import io.itmca.lifepuzzle.domain.user.repository.UserRepository;
 import io.itmca.lifepuzzle.global.constants.ServerConstant;
 import io.itmca.lifepuzzle.global.exception.UserNotAccessibleToHeroException;
 import io.itmca.lifepuzzle.global.exception.UserNotShareHeroAuthException;
+import io.itmca.lifepuzzle.global.file.domain.UserProfileImage;
 import io.itmca.lifepuzzle.global.file.service.S3UploadService;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
