+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/global/file/FileDimension.java
@@ -1,4 +1,4 @@
-package io.itmca.lifepuzzle.global.model;
+package io.itmca.lifepuzzle.global.file;
 
 public record FileDimension(int width, int height) {
 }
