+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/global/file/domain/StoryFile.java
@@ -1,4 +1,4 @@
-package io.itmca.lifepuzzle.domain.content.file;
+package io.itmca.lifepuzzle.global.file.domain;
 
 import java.util.List;
 import lombok.Builder;
