+++ services/lifepuzzle-api/src/test/java/io/itmca/lifepuzzle/domain/hero/entity/HeroTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.navercorp.fixturemonkey.FixtureMonkey;
 import com.navercorp.fixturemonkey.api.introspector.ConstructorPropertiesArbitraryIntrospector;
-import io.itmca.lifepuzzle.domain.hero.file.HeroProfileImage;
+import io.itmca.lifepuzzle.global.file.domain.HeroProfileImage;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
